Refactor training script into package modules
