+++ botcopier/training/pipeline.py
@@ -8,13 +8,15 @@ import gzip
 import importlib.metadata as importlib_metadata
 import json
 import logging
+import os
 import shutil
 import time
 from contextlib import nullcontext
 from pathlib import Path
-from typing import Iterable, Sequence
+from typing import Callable, Iterable, Sequence
 
 import numpy as np
+import optuna
 import pandas as pd
 import psutil
 from joblib import Memory
@@ -176,6 +178,109 @@ def _write_dependency_snapshot(out_dir: Path) -> Path:
     dep_path = out_dir / "dependencies.txt"
     dep_path.write_text("\n".join(packages))
     return dep_path
+
+
+def _trial_logger(
+    csv_path: Path,
+) -> Callable[[optuna.study.Study, optuna.trial.FrozenTrial], None]:
+    """Return a callback that appends trial information to ``csv_path``."""
+
+    def _callback(study: optuna.study.Study, trial: optuna.trial.FrozenTrial) -> None:
+        row = {
+            "trial": trial.number,
+            **trial.params,
+            "seed": trial.user_attrs.get("seed"),
+            "profit": trial.values[0],
+            "sharpe": trial.values[1],
+            "max_drawdown": trial.values[2],
+        }
+        df = pd.DataFrame([row])
+        df.to_csv(csv_path, mode="a", header=not csv_path.exists(), index=False)
+
+    return _callback
+
+
+def _objective_factory(
+    max_drawdown: float | None, var_limit: float | None
+) -> Callable[[optuna.trial.Trial], tuple[float, float, float]]:
+    """Return a multi-objective that includes risk penalties."""
+
+    def _objective(trial: optuna.trial.Trial) -> tuple[float, float, float]:
+        seed = trial.suggest_int("seed", 0, 9999)
+        trial.set_user_attr("seed", seed)
+        x = trial.suggest_float("x", -10.0, 10.0)
+        rng = np.random.default_rng(seed)
+        noise = rng.normal()
+        risk = abs(x) / 10.0
+        trial.set_user_attr("max_drawdown", risk)
+        trial.set_user_attr("var_95", risk)
+        profit = -((x - 2) ** 2) + noise
+        if max_drawdown is not None and risk > max_drawdown:
+            profit -= risk - max_drawdown
+        if var_limit is not None and risk > var_limit:
+            profit -= risk - var_limit
+        sharpe = profit / (risk + 1e-6)
+        return profit, sharpe, risk
+
+    return _objective
+
+
+def run_optuna(
+    n_trials: int = 10,
+    csv_path: Path | str = "hyperparams.csv",
+    model_json_path: Path | str = "model.json",
+    *,
+    max_drawdown: float | None = None,
+    var_limit: float | None = None,
+) -> optuna.study.Study:
+    """Run a small Optuna study and record trial information."""
+
+    csv_path = Path(csv_path)
+    model_json_path = Path(model_json_path)
+
+    sampler = optuna.samplers.RandomSampler(seed=0)
+    study = optuna.create_study(
+        directions=["maximize", "maximize", "minimize"], sampler=sampler
+    )
+    objective = _objective_factory(max_drawdown, var_limit)
+    study.optimize(objective, n_trials=n_trials, callbacks=[_trial_logger(csv_path)])
+
+    def _select_trial() -> optuna.trial.FrozenTrial:
+        candidates = [t for t in study.best_trials]
+        if max_drawdown is not None:
+            candidates = [t for t in candidates if t.values[2] <= max_drawdown]
+        if var_limit is not None:
+            candidates = [
+                t
+                for t in candidates
+                if t.user_attrs.get("var_95", t.values[2]) <= var_limit
+            ]
+        if not candidates:
+            candidates = list(study.best_trials)
+        return max(candidates, key=lambda t: (t.values[0], t.values[1]))
+
+    best = _select_trial()
+    relative_csv = os.path.relpath(csv_path, model_json_path.parent)
+    risk = best.user_attrs.get("max_drawdown", 0.0)
+    model_data = {
+        "metadata": {
+            "hyperparam_log": relative_csv,
+            "selected_trial": {
+                "number": best.number,
+                "profit": best.values[0],
+                "sharpe": best.values[1],
+                "max_drawdown": best.values[2],
+            },
+        },
+        "risk_params": {"max_drawdown": max_drawdown, "var_limit": var_limit},
+        "risk_metrics": {
+            "max_drawdown": risk,
+            "var_95": best.user_attrs.get("var_95", risk),
+        },
+    }
+    model_json_path.write_text(json.dumps(model_data))
+
+    return study
 
 
 def train(
